Gate guest/boot SSH bring-up behind a nosshd build tag
Split guest/boot/boot.go so the SSH-specific portion lives in two
build-tag-gated files:

  guest/boot/ssh_enabled.go   //go:build linux && !nosshd
  guest/boot/ssh_disabled.go  //go:build linux && nosshd

The default build keeps SSH on — current behavior is preserved and no
consumer changes are required. Pass -tags=nosshd at build time to swap
in the stub, which strips golang.org/x/crypto/ssh and guest/sshd from
the link-time graph entirely.

Semantics:
- Default build (no tag): SSH compiled in, runtime behavior identical
  to before this change.
- -tags=nosshd: SSH stripped. Callers MUST opt out at runtime via
  WithoutSSH; otherwise boot.Run returns the new sentinel
  ErrSSHNotBuilt and fails loud rather than silently boot without SSH.

The orchestrator in boot.go no longer imports golang.org/x/crypto/ssh
or guest/sshd directly. It calls a single
bringUpSSH(logger, cfg, envVars) provided by whichever of the two
gated files is compiled in. ParseAuthorizedKeys moves into
ssh_enabled.go (its return type references ssh.PublicKey).

Tests split by the same tag:
- boot_test.go (both builds): option / config tests with no SSH types.
- ssh_enabled_test.go (//go:build !nosshd): ParseAuthorizedKeys tests
  plus the generateTestKey helper that depends on x/crypto/ssh.
- ssh_disabled_test.go (//go:build nosshd): verifies the stub's two
  documented behaviors — no-op when WithoutSSH is set, ErrSSHNotBuilt
  when it is not.

Taskfile: the test, test-nocgo, and build-nocgo targets now run both
the default and -tags=nosshd variants so CI exercises both code paths.

The host-side microvm.WithoutSSH doc and the guest/boot package doc
both mention ErrSSHNotBuilt so callers know what to expect when they
forget to either drop the tag or set WithoutSSH.

Verification:
  go list -deps ./guest/boot/                  -> includes crypto/ssh
  go list -tags=nosshd -deps ./guest/boot/     -> no crypto/ssh
